Refactor with some critical abstractions
